generator: Dispatch gen() using node type introspection.
It is possible to determine the node type at runtime, so therefore it
seems unnecessary to specify the node type in the call to gen().  Just
like before, if this option is used the template will be fetched from a
predefined default_templates {} variable.

If a second argument is given, it can now only be the name of a
template, when it is desired to explicitly specify the template
(overriding the default, if it was defined).

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>