Run CI test with python 3.8 instead of 3.9
It was reported that 3.8.5 is the python version in latest Ubuntu or
latest Ubuntu-LTS.  With a recent fix the code now works with 3.8, so
this commit suggests that future changes should also be tested
to be compatible with this version.

Noteworthy is that the GitHub action we use has no python 3.8.5
available - the closest one is 3.8.11, which is also what we get when we
request 3.8 as in this commit. Hopefully this is close enough for a
compatibility check.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>