Bugfix AnyTree path (wrong parent on child nodes)
Ensure nodes have the right parent node so that the AnyTree hierarchy
path is correct.  Although it didn't affect most generators, this means
AST can be navigated either through navigating the member variables of
the AST objects, or by using AnyTree methods to address/search/etc. in the
hierarchical path that AnyTree keeps track of.