vsc_parser: Remove all unused functions
All the previous code in this file, to read YAML into an AST tree, is
now superseded by vsc_ast.py.

There are a few files and scripts referring to this file so at the
moment it remains as a wrapper around vst_ast but soon it should
probably be replaced completely.