Create helper scripts and dev container
This creates a package of the recommended environment to use as an
interactive dev-container.  But it can also be used as a
command-container that as a replacement for any of the tool executables.

The main advantage is that a containerized distribution of the tool
does not require installing the IFEX project or any other python
prerequisites onto the host system.

The provided Dockerfile installs most things during the image build,
for quicker turnaround once the docker image has been built and cached.

The scripts/ directory now contains a combination of helper scripts, many of
which are used by the Dockerfile to set up the environment but they can
be useful directly on any host as well.

Signed-off-by: Gunnar Andersson <gunnar_dev@nospam@novaspring.eu>