setup.py: Packaging all template files, new structure
Templates are not only limited to extention .tpl and .html so we
package all file types.

At the same time, templates can also now be stored in sub-directories of
templates dir. (Only in the sub-directories).  In other words, each
generation case uses its own sub-directory to define its templates.

If we only package files in the subdirectories and not the main template
directory then we avoid getting __init__.py files into the mix.

Signed-off-by: Gunnar Andersson <gunnar_dev@nospam@novaspring.eu>