Introduce JinjaTemplateEnv module
This module wraps the functions and data dealing with the jinja2
environment.  It is currently a single-instance module but the purpose
is to enable instantiating multiple environments.  In addition to
jinja2 standard features it supports the ifex_generator idea of the
default_templates for a particular node type, etc.

This refactoring is done step-by-step to ensure that unit tests are
still running OK through each step.

Signed-off-by: Gunnar Andersson <gunnar_dev@nospam@novaspring.eu>