Programs to "sort" and compare IFEX content
stable_sort_ifex defines a known order of keys so that IFEX files can be
reliably compared.  It is a module but can also be run as a standalone
script.

The diff_ifex.py script will print the diff between two files, after
normalizing the order of the content using the stable_sort_ifex module.
It can also be used with an external diff program.

Signed-off-by: Gunnar Andersson <gunnar_dev@nospam@novaspring.eu>