pyenv scripts: set up path to pyenv correctly
These scripts are used to set up a working environment - primarily when
creating a container with the working environment but might be useful
also on a "native" setup.

The installation script now inserts the pyenv init (set up path etc.)
in ~/.bashrc.

But in the  install_python_version_in_pyenv.sh script we can set up PATH
to the pyenv binary directly, before using it.

Signed-off-by: Gunnar Andersson <gunnar_dev@nospam@novaspring.eu>