ifex_ast: Descriptions s/primitive/fundamental/
In the documentation the built-in types are called fundamental
so the detailed descriptions should do the same.

A "primitive" type can be considered int/float/string/boolean
but also a few others like set and map are "built-in" into IFEX, i.e.
they are also fundamental.

(Non-fundamental types are enumerations, typedefs and structs.)

Signed-off-by: Gunnar Andersson <gunnar_dev@nospam@novaspring.eu>