ifex_ast: Allow enumerations to have any type.
While most usage of enumerations will be backed by an integer type,
there is no real reason to disallow other types.
For example, enumerations where each value is a string-constant.
From now on the limitation is described as the backing type must
resolve to a primitive type (e.g. it can't be a Struct, or array).

This is a big change to the specification but since it is "growing"
the set of possibilities it might be considered backwards compatible.
(MINOR)

The field that has "typing.Any" type annotation is now skipped over in
the document generation (ifex_ast_doc.py) since it does not refer to
any AST class. While doing that there was some small refactoring to
simplify the logic also.

Signed-off-by: Gunnar Andersson <gunnar_dev@nospam@novaspring.eu>