Remove byteBuffer type (use array of bytes)
It wasn't really implemented in several places anyway, but this
removes some remaining references to this type.

The byteBuffer type is superfluous since it can be represented
by an array of uint8.  It was at an earlier time also removed from the
VSS project.  Another alternative is the IFEX 'opaque' type. It was
conceived to indicate that something in reality has structure but that
it cannot (or is chosen not to) have a detailed represented in the IFEX
description.  An Opaque type is presumably also transferred as "just a
bunch of bytes" in a computing system - only the name infers there is
some additional meaning.

Signed-off-by: Gunnar Andersson <gunnar_dev@nospam@novaspring.eu>