Add regression test (diff), Github workflow and scripts
This runs through most outputs we can find (various generators
and documentation generation) and reports any differences compared
to previous versions.  This is useful for spotting regressions and
being confident about refactorings.

At this time, comparison of the result must be done manually -
the workflow does not determine "failure" on a regression (but that is
of course done by the normal unit tests, etc.).

NOTE:  This removes other workflows from regression test branch
because they are superfluous in the current usage (they will be run
when a PR is sent, etc.)

Later, if the regression test is introduced in the PR flow, some
conditions could be set up to make this better.

Signed-off-by: Gunnar Andersson <gunnar_dev@nospam@novaspring.eu>