Bugfix docs link to VSS interoperability
Github pages have some ideosyncracies.  If a link points to "file.md"
then GitHub seems to render the raw file (not even interpreting
Markdown).

In order to get the nicely-generated HTML pages, and stay within the
GitHub pages, we need to link to simply: "file"

I believe changing this instead breaks the link if navigating from
within the markdown document inside the code base. :-(  But we
prioritize the GitHub pages for documentation so this has to be done!

Signed-off-by: Gunnar Andersson <gunnar.andersson@mercedes-benz.com>