Workflows: Add automated Release of JSON schema
This adds a GitHub action that implements automated deployment (to the
GitHub Releases section) for the JSON schema file.

It is deployed to the release section because some projects want to
depend on the file as a specification of the IFEX Core IDL.

But since the JSON schema is a generated artifact based on the internal
model defined in the source code (ifex_ast.py), it should not be
committed to the source code repository, since there can be a risk of
mismatch.  Therefore it is generated.

Note: It is set up to create a release, and a tag, by inserting
a line like this in the commit:

Signed-off-by: Gunnar Andersson <gunnar.andersson@mercedes-benz.com>