JSON schema generation: Add a description/version field
Custom fields are allowed in the json schema but here we use the
semi-standard description field.  We don't intend to repeat a long
description of what IFEX Core IDL is here, but it can be used to
indicate the schema version, which is defined using the release tag.

It will be placed there by replacing TAG-PLACEHOLDER in CI.

RELEASE-TAG: spec_v1.3

Signed-off-by: Gunnar Andersson <gunnar.andersson@mercedes-benz.com>