D-Bus / sdbus-c++: cmake flag was changed
The helper build script that compiles the sdbus-c++-xml2cpp compiler
is based on git-cloning the latest master branch, instead of locking
it into a specific version.  An incompatible change was introduced
there, for the name of the CMAKE flag that is used to cause the
code-generator binary to be built.  Renaming the flag in our script
makes the build work again.

(It was, and still is, *NOT* caught by regression testing, since this is
a small and complementary helper, and not really part of the core
delivered tools, so not a big issue.)

Signed-off-by: Gunnar Andersson <gunnar.andersson@mercedes-benz.com>