Implement merge-model functions
This is a fundamental implementation that implements/defines the rules
for combining Layers.  It implements the ability to add, delete, and
change information, both in the case of overlays that modify existing
interface definitions within the IFEX Core IDL scope, and new layers
that augment the information with new information types.

In addition to combining layers it is useful in model-to-model
translators where you may need to build up different parts of the
resulting AST separately and then put them together.

Signed-off-by: Gunnar Andersson <gunnar.andersson@mercedes-benz.com>