protobuf_to_ifex: Robust way to get grammar file
Previous script used current working directory as starting point but that
doesn't work when this module is imported from another script in another
location. This change is intended to guarantee that we find the grammar file
relative to the module directory (also if it is an import from another
main file).

Signed-off-by: Gunnar Andersson <gunnar.andersson@mercedes-benz.com>