Add output templates for Protobuf/gRPC (based on parsed Proto AST)
These output Jinja templates define an output from the gRPC/protobuf AST model
to text.  Previously, a rudimentary template existed going directly from IFEX
AST model to Protobuf text.  This new code is however based on the
*Protobuf* AST as input to the Jinja templates.

This prepares for a IFEX-to-gRPC model-to-model transformation,
followed by a subsequent textual printout, as opposed to going from
IFEX-model directly to foreign text format.

The included test program grpc_print.py can be used to parse an existing
protobuf/gRPC description, and then print it out again from the resulting AST.

The printout should not be expected to perfectly identical with the
input because the order of items may change, flexibility in the syntax,
and whitespace.  However, running the script twice (re-parsing the
output from the first time) should, if it functions correctly, yield an
identical output the second time.

Such a test won't notice if something was lost in either stage, such as
an unsupported feature, but it will at least verify that the output is
valid proto syntax, according to our own parser, at least.

Signed-off-by: Gunnar Andersson <gunnar.andersson@mercedes-benz.com>