rule_translator: Implement Delegate feature
This feature allows the translation table to capture a value
from one input node, and assign it to a later feature node.

It delegates the processing of input features at one level of
the tree to a later (typically deeper nested) node type.

This can handle some of the situations where the same type of
information is defined at different levels of the language hierarchy
in the incoming vs. outgoing language.

Signed-off-by: Gunnar Andersson <gunnar.andersson@mercedes-benz.com>