rule_translator: Don't append None objects to list
If attribute in target object is a list, we normally append
to it, but if the new value is None, we don't want it appended.

This seems to remove some issue, but there is a small risk doing this
will instead hide other bugs (None value should not appear).  Let's see
how it works out.

Signed-off-by: Gunnar Andersson <gunnar.andersson@mercedes-benz.com>