Move IFEX-specific AST utils to common utils
Utility functions previously in IFEX-specific files are actually generic
and can be used for multiple similar ASTs.  The common/ast_utils file is
the current place to bring them together.

After this, the ifex_ast_construction file is obsolete -> deleted
and the type_checking_constructor_mixin is moved to common area.

Signed-off-by: Gunnar Andersson <gunnar.andersson@mercedes-benz.com>