Remove weird junk from setup.py (TemplateDir)
The special stuff caused issues during refactoring and changing
imports between files - the whole project then refused to install
due to some circular-dependencies or other import issues caused
by this in the setup.py

I believe we don't need the special handling of the templatedir anymore.
If we do, it should be done differently anyway.

Signed-off-by: Gunnar Andersson <gunnar.andersson@mercedes-benz.com>
Signed-off-by: Marco Benelli <marco.benelli@mercedes-benz.com>