protobuf: Process enumerations separately
In grpc/protobuf IDL, enumerations can be defined inside of messages.
We primarily treat each message as a datatype definition and therefore
convert them to a Struct type.

A type-definition inside, like an enumeration, does not belong there in
the IFEX way of modeling.  It cannot be mapped into something equivalent
inside of a Struct, so the enumeration definitions are instead collected
separately and placed in an appropriate location for the IFEX model.

Signed-off-by: Gunnar Andersson <gunnar.andersson@mercedes-benz.com>
Signed-off-by: Marco Benelli <marco.benelli@mercedes-benz.com>