rule_translator: Set unsupported feature to Warning
During development we need to process files that are parsed correctly
but not supported for translation in a certain case.

If the log level is unspecified, the default is ERROR. It is better if
the program stays quiet about this in development.  During actual use,
it is possible to set a lower threshold to get notified of that log
level..

Signed-off-by: Gunnar Andersson <gunnar.andersson@mercedes-benz.com>
Signed-off-by: Marco Benelli <marco.benelli@mercedes-benz.com>