Defined Layer types: ValueConstraint and Timing
The structure for Layer Types is introduced, and two common/important
layer types are specified.  They serve as examples but are also intended
as a real proposal for standardized layer types.

For each layer type, a JSON schema is provided to identify and validate
input files against the layer format.  The READMEs in the main layer types
directory and for each layer type describe how they can be used.

Signed-off-by: Gunnar Andersson <gunnar.andersson@mercedes-benz.com>
Signed-off-by: Marco Benelli <marco.benelli@mercedes-benz.com>